backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain_community
langchain_core
langchain_groq
groq
python-dotenv